# Optimize no-op effects; 11% overall perf win