Various functions call ModdedDex#getEffect with null, undefined, or an empty string. For example, in the common case, Battle#getWeather calls getEffect with an empty string. In these cases, getEffect basically returns the same object each time: a PureEffect indicating that no effect with the given name exists.

Creating the dummy PureEffect is expensive and puts pressure on the JavaScript garbage collector. Memoize the creation in getEffect to improve performance.

(As a bonus, special-casing null, undefined, and empty strings also lets us drop some noisy truthiness checks in getEffect.)

On my machine, this commit speeds up Pokemon Showdown's tests by 11%.

Methodology: With and without this commit, I ran mocha four times with zsh' 'time' builtin, dropped the first result, and averaged the wall times:

    mocha times before this commit:
    18.20s user 0.33s system 118% cpu 15.704 total
    17.91s user 0.34s system 118% cpu 15.454 total
    18.11s user 0.33s system 118% cpu 15.558 total

    mocha times after this commit:
    16.53s user 0.35s system 120% cpu 13.994 total
    16.43s user 0.34s system 120% cpu 13.918 total
    16.54s user 0.32s system 120% cpu 14.009 total

    Hardware:
    Mid 2012 MacBook Pro
    2.6 GHz Intel Core i7

    Software:
    Node v9.0.0
    macOS 10.10.5